pyproject.toml
[project]
name = "music-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py[voice]>=2.6.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "spotipy>=2.25.2",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "types-requests>=2.32.4.20260107",
]

[tool.mypy]
cache_dir = ".venv/mypy_cache"
